requirements.txt
Pillow
requests
torch
torchvision